Move to a more standard filenaming scheme and get tests running in non-windows environments.
