+++ X12.gemspec
@@ -23,16 +23,16 @@ Gem::Specification.new do |s|
     "Rakefile",
     "TODO",
     "VERSION",
-    "lib/X12.rb",
-    "lib/X12/Base.rb",
-    "lib/X12/Composite.rb",
-    "lib/X12/Empty.rb",
-    "lib/X12/Field.rb",
-    "lib/X12/Loop.rb",
-    "lib/X12/Parser.rb",
-    "lib/X12/Segment.rb",
-    "lib/X12/Table.rb",
-    "lib/X12/XMLDefinitions.rb",
+    "lib/x12.rb",
+    "lib/x12/base.rb",
+    "lib/x12/composite.rb",
+    "lib/x12/empty.rb",
+    "lib/x12/field.rb",
+    "lib/x12/loop.rb",
+    "lib/x12/parser.rb",
+    "lib/x12/segment.rb",
+    "lib/x12/table.rb",
+    "lib/x12/xml_definitions.rb",
     "test/tc_factory_270.rb",
     "test/tc_factory_270interchange.rb",
     "test/tc_factory_997.rb",
