+++ lib/x12/base.rb
