+++ lib/x12/composite.rb
