+++ lib/x12/empty.rb
