+++ lib/x12/field.rb
