+++ lib/x12/loop.rb
