+++ lib/x12/parser.rb
