+++ lib/x12/segment.rb
