+++ lib/x12/table.rb
