+++ lib/x12/xml_definitions.rb
